Refactor monitoree messaging via sms and call
