+++ app/errors/invalid_messaging_method_error.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class InvalidMessagingMethodError < StandardError
+  def initialize(klass, method)
+    super("#{klass}##{method} is not a valid message method")
+  end
+end
