+++ config/environments/development.rb
@@ -1,5 +1,6 @@
 require "active_support/core_ext/integer/time"
 
+Rails.application.routes.default_url_options = { host: 'localhost', port: 3000 }
 Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
