+++ config/environments/test.rb
@@ -4,7 +4,7 @@ require "active_support/core_ext/integer/time"
 # test suite. You never need to work with it otherwise. Remember that
 # your test database is "scratch space" for the test suite and is wiped
 # and recreated between test runs. Don't rely on the data there!
-
+Rails.application.routes.default_url_options = { host: 'localhost', port: 3000 }
 Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
 
@@ -48,7 +48,6 @@ Rails.application.configure do
   # The :test delivery method accumulates sent emails in the
   # ActionMailer::Base.deliveries array.
   config.action_mailer.delivery_method = :test
-  config.action_mailer.default_url_options = { :host => "saraalert-test.mitre.org" }
   # Print deprecation notices to the stderr.
   config.active_support.deprecation = :stderr
